requirements.txt
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blessed==1.19.1
certifi==2022.9.24
charset-normalizer==2.1.1
crispy-bootstrap5==0.6
Django==4.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-grappelli==3.0.3
django-livereload==1.7
django-livereload-server==0.4
django-picklefield==3.1
django-q==1.3.9
idna==3.4
mysqlclient==2.1.1
pbr==5.10.0
python-dateutil==2.8.2
redis==3.5.3
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
testresources==2.0.1
tornado==6.2
urllib3==1.26.12
wcwidth==0.2.5
